fastapi/requirements.txt
# FastAPI 服务依赖
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.11.3

# LangChain 相关库
langchain==0.3.28
langchain-openai==0.3.35
langchain-community==0.3.21
langchain-deepseek==0.1.3
langchain-classic==1.0.3
langgraph==0.3.34
langchain-core==0.3.84

# 数据库
pymysql==1.1.1
sqlalchemy==2.0.40

# 安全与加密
bcrypt==4.3.0
python-jose[cryptography]==3.4.0

# 环境变量
python-dotenv==1.1.0

# Docker SDK
docker==7.1.0

# 数据处理
pandas==2.2.3
numpy==2.2.6

# 可视化
pyecharts==2.0.8

# HTTP 请求
requests==2.28.1
urllib3==1.26.20
